# apps/api/src/rag_forge/api/routes.py
"""HTTP routes, grouped by capability. Split into modules once a group outgrows a screen."""

from __future__ import annotations

from datetime import UTC, datetime
from typing import Any

from fastapi import APIRouter, HTTPException, Request, status

from rag_forge import __version__
from rag_forge.api.schemas import (
    ComponentHealth,
    ComponentState,
    EnvironmentResponse,
    HealthStatus,
    NotImplementedDetail,
    RagComponents,
    RetrievalEvaluationRequest,
    RetrievalEvaluationResponse,
    RouterRequest,
    RouterResponse,
    VerifierDescriptor,
)
from rag_forge.domain.models import (
    EvidenceParams,
    GenerationParams,
    Metric,
    RetrievalMode,
    RetrievalRequest,
)
from rag_forge.evaluation import retrieval_metrics as rm
from rag_forge.ingestion.service import CorpusNotFoundError
from rag_forge.provenance.environment import capture_environment
from rag_forge.rag.service import RagService
from rag_forge.retrieval.rerank import Reranker
from rag_forge.retrieval.service import CorpusVersionNotFoundError, RetrievalService
from rag_forge.router.service import RouterComponentNotAvailableError
from rag_forge.storage.base import CorpusStore

router = APIRouter(prefix="/api/v1")

NOT_IMPLEMENTED: dict[int | str, dict[str, Any]] = {501: {"model": NotImplementedDetail}}


def _store(request: Request) -> CorpusStore:
    store: CorpusStore = request.app.state.store
    return store


# --- System -------------------------------------------------------------------


@router.get("/health", response_model=HealthStatus, tags=["system"])
def health(request: Request) -> HealthStatus:
    started_at: datetime = request.app.state.started_at
    store = _store(request)
    components = [
        ComponentHealth(name="api", state=ComponentState.OK, detail=f"rag-forge {__version__}"),
        ComponentHealth(
            name="metadata_store",
            state=ComponentState.OK,
            detail=f"{store.kind}: {getattr(store, 'path', '')}",
        ),
        ComponentHealth(
            name="vector_index", state=ComponentState.NOT_CONFIGURED, detail="no index backend"
        ),
        _generator_health(request.app.state.rag),
        _reranker_health(request.app.state.reranker),
    ]
    healthy = all(c.state != ComponentState.ERROR for c in components)
    return HealthStatus(
        status="ok" if healthy else "degraded",
        version=__version__,
        started_at=started_at,
        uptime_seconds=(datetime.now(UTC) - started_at).total_seconds(),
        components=components,
    )


def _reranker_health(reranker: Reranker) -> ComponentHealth:
    """Configured, not loaded: health never triggers a model download."""
    spec = reranker.spec
    return ComponentHealth(
        name="reranker", state=ComponentState.OK, detail=f"{spec.model}@{spec.revision[:8]}"
    )


def _generator_health(rag: RagService) -> ComponentHealth:
    """Configured, not loaded: health never triggers a model download."""
    d = rag.generators[rag.default_generator].describe()
    revision = f"@{d.revision[:8]}" if d.revision else ""
    state = "loaded" if d.loaded else "loads on first answer"
    return ComponentHealth(
        name="llm_provider",
        state=ComponentState.OK,
        detail=f"{d.model}{revision} ({d.provider}, {state})",
    )


@router.get("/provenance/environment", response_model=EnvironmentResponse, tags=["provenance"])
def environment() -> EnvironmentResponse:
    return EnvironmentResponse(environment=capture_environment())


# --- Routing -------------------------------------------------------


@router.post(
    "/router/decide", response_model=RouterResponse, tags=["router"], responses=NOT_IMPLEMENTED
)
def decide(body: RouterRequest, request: Request) -> RouterResponse:
    """Analyse a query and decide its retrieval configuration without retrieving."""
    service: RetrievalService = request.app.state.retrieval
    try:
        req = RetrievalRequest(
            query=body.query,
            top_k=body.top_k,
            version=body.version,
            mode=RetrievalMode.ADAPTIVE,
            router=body.router,
        )
        query, routed, routing = service.route(body.corpus_id, req)
    except ValueError as exc:  # pydantic validation of the derived request (e.g. blank query)
        raise HTTPException(status.HTTP_422_UNPROCESSABLE_CONTENT, detail=str(exc)) from exc
    except (CorpusNotFoundError, CorpusVersionNotFoundError) as exc:
        raise HTTPException(status.HTTP_404_NOT_FOUND, detail=str(exc)) from exc
    except RouterComponentNotAvailableError as exc:
        detail = NotImplementedDetail(capability="Router", message=str(exc))
        raise HTTPException(status.HTTP_501_NOT_IMPLEMENTED, detail=detail.model_dump()) from exc
    return RouterResponse(query=query, request=routed, routing=routing)


# --- RAG -----------------------------------------------------------------------


@router.get("/rag/components", response_model=RagComponents, tags=["rag"])
def rag_components(request: Request) -> RagComponents:
    rag: RagService = request.app.state.rag
    return RagComponents(
        generators=[g.describe() for g in rag.generators.values()],
        default_generator=rag.default_generator,
        verifiers=[
            VerifierDescriptor(
                name=v.name,
                version=v.version,
                detects_contradiction=v.detects_contradiction,
                thresholds=v.thresholds(),
                config_hash=v.config_hash(),
            )
            for v in rag.verifiers.values()
        ],
        prompt_template=rag.template.id,
        evidence_defaults=EvidenceParams(),
        generation_defaults=GenerationParams(generator=rag.default_generator),
    )


# --- Evaluation ---------------------------------------------------------------


@router.post(
    "/evaluation/retrieval", response_model=RetrievalEvaluationResponse, tags=["evaluation"]
)
def evaluate_retrieval(body: RetrievalEvaluationRequest) -> RetrievalEvaluationResponse:
    ranked, rel = body.ranked_ids, body.relevance
    try:
        metrics = [Metric(name="mrr", value=rm.reciprocal_rank(ranked, rel))]
        for k in sorted(set(body.ks)):
            metrics += [
                Metric(name="recall", k=k, value=rm.recall_at_k(ranked, rel, k)),
                Metric(name="precision", k=k, value=rm.precision_at_k(ranked, rel, k)),
                Metric(name="ndcg", k=k, value=rm.ndcg_at_k(ranked, rel, k)),
            ]
    except ValueError as exc:
        raise HTTPException(status.HTTP_422_UNPROCESSABLE_CONTENT, detail=str(exc)) from exc
    return RetrievalEvaluationResponse(metrics=metrics)