# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RAG FORGE API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/router/decide": {
      "post": {
        "tags": [
          "router"
        ],
        "summary": "Decide",
        "description": "Analyse a query and decide its retrieval configuration without retrieving.",
        "operationId": "decide_api_v1_router_decide_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouterResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/evaluation/retrieval": {
      "post": {
        "tags": [
          "evaluation"
        ],
        "summary": "Evaluate Retrieval",
        "operationId": "evaluate_retrieval_api_v1_evaluation_retrieval_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrievalEvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrievalEvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Bm25Params-Output": {
        "properties": {
          "k1": {
            "type": "number",
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "K1",
            "description": "Term-frequency saturation",
            "default": 1.2
          },
          "b": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "B",
            "description": "Chunk-length normalisation",
            "default": 0.75
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "k1",
          "b"
        ],
        "title": "Bm25Params"
      },
...
      "ContentOrigin": {
        "type": "string",
        "enum": [
          "retrieved",
          "inferred",
          "generated",
          "measured",
          "simulated"
        ],
        "title": "ContentOrigin",
        "description": "Where a piece of information came from. Every value shown to a user carries one."
      },
...
      "FusionMethod": {
        "type": "string",
        "enum": [
          "rrf",
          "weighted"
        ],
        "title": "FusionMethod"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HybridParams-Output": {
        "properties": {
          "retrievers": {
            "items": {
              "$ref": "#/components/schemas/RetrievalStrategy"
            },
            "type": "array",
            "title": "Retrievers",
            "description": "Component strategies to fuse"
          },
          "fusion": {
            "$ref": "#/components/schemas/FusionMethod",
            "default": "rrf"
          },
          "rrf_k": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Rrf K",
            "description": "RRF smoothing constant",
            "default": 60
          },
          "weights": {
            "additionalProperties": {
              "type": "number"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/RetrievalStrategy"
            },
            "type": "object",
            "title": "Weights",
            "description": "Weighted fusion only: one weight per component, each 0..1, summing to 1"
          },
          "candidate_k": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Candidate K",
            "description": "Candidates requested from each component",
            "default": 50
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "retrievers",
          "fusion",
          "rrf_k",
          "weights",
          "candidate_k"
        ],
        "title": "HybridParams"
      },
...
      "Metric": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "K"
          },
          "origin": {
            "$ref": "#/components/schemas/ContentOrigin",
            "default": "measured"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "value",
          "k",
          "origin"
        ],
        "title": "Metric"
      },
...
      "NotImplementedDetail": {
        "properties": {
          "capability": {
            "type": "string",
            "title": "Capability"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "capability",
          "message"
        ],
        "title": "NotImplementedDetail"
      },
...
      "Query": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "text",
          "metadata",
          "created_at"
        ],
        "title": "Query"
      },
      "QueryAnalysis": {
        "properties": {
          "analyzer": {
            "type": "string",
            "title": "Analyzer"
          },
          "analyzer_version": {
            "type": "string",
            "title": "Analyzer Version"
          },
          "config_hash": {
            "type": "string",
            "title": "Config Hash",
            "description": "Hash of the analyzer's weights and thresholds"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "normalized_query": {
            "type": "string",
            "title": "Normalized Query"
          },
          "features": {
            "$ref": "#/components/schemas/QueryFeatures"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/QuerySignal"
            },
            "type": "array",
            "title": "Signals"
          },
          "labels": {
            "$ref": "#/components/schemas/QueryLabels"
          },
          "corpus": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CorpusQuerySignals"
              },
              {
                "type": "null"
              }
            ]
          },
          "analysis_hash": {
            "type": "string",
            "title": "Analysis Hash",
            "description": "Hash of everything above; equal inputs, equal hash"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "analyzer",
          "analyzer_version",
          "config_hash",
          "query",
          "normalized_query",
          "features",
          "signals",
          "labels",
          "corpus",
          "analysis_hash"
        ],
        "title": "QueryAnalysis"
      },
...
      "QueryFeatures": {
        "properties": {
          "char_count": {
            "type": "integer",
            "title": "Char Count"
          },
          "token_count": {
            "type": "integer",
            "title": "Token Count",
            "description": "Word tokens (Unicode word characters)"
          },
          "bm25_terms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Bm25 Terms",
            "description": "Distinct terms BM25 searches (its analyzer)"
          },
          "key_terms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Terms",
            "description": "bm25_terms that are not function words"
          },
          "function_word_ratio": {
            "type": "number",
            "title": "Function Word Ratio",
            "description": "Share of word tokens that are function words (stopwords, interrogatives, auxiliaries, pronouns, prepositions)"
          },
          "is_question": {
            "type": "boolean",
            "title": "Is Question",
            "description": "Ends with '?' or starts with a question/request word"
          },
          "question_word": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Word",
            "description": "The leading question or request word"
          },
          "question_type": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "quoted_phrases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Quoted Phrases"
          },
          "identifiers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Identifiers",
            "description": "Tokens with digits, inner capitals, underscores, dots or 2+ capitals"
          },
          "capitalized_terms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capitalized Terms",
            "description": "Capitalised words not at sentence start"
          },
          "numbers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Numbers"
          },
          "entities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Entities",
            "description": "Quoted phrases, identifiers and capitalised terms"
          },
          "concept_segments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Concept Segments",
            "description": "Parts split on and/or/vs/commas/semicolons that contain a key term"
          },
          "comparison_markers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Comparison Markers"
          },
          "multi_hop_markers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Multi Hop Markers"
          },
          "temporal_markers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Temporal Markers"
          },
          "negation_markers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Negation Markers"
          },
          "ambiguity_markers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ambiguity Markers"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "char_count",
          "token_count",
          "bm25_terms",
          "key_terms",
          "function_word_ratio",
          "is_question",
          "question_word",
          "question_type",
          "quoted_phrases",
          "identifiers",
          "capitalized_terms",
          "numbers",
          "entities",
          "concept_segments",
          "comparison_markers",
          "multi_hop_markers",
          "temporal_markers",
          "negation_markers",
          "ambiguity_markers"
        ],
        "title": "QueryFeatures",
        "description": "Measured properties of the query text. Every list keeps first-occurrence order."
      },
...
      "QuestionType": {
        "type": "string",
        "enum": [
          "definition",
          "procedural",
          "explanatory",
          "comparison",
          "list",
          "boolean",
          "factoid",
          "keyword"
        ],
        "title": "QuestionType"
      },
...
      "RerankParams-Output": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "A registered reranker model",
            "default": "cross-encoder/ms-marco-MiniLM-L-6-v2"
          },
          "candidate_k": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Candidate K",
            "description": "Upstream candidates the reranker scores",
            "default": 50
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "enabled",
          "model",
          "candidate_k"
        ],
        "title": "RerankParams"
      },
...
      "RetrievalEvaluationRequest": {
        "properties": {
          "ranked_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ranked Ids"
          },
          "relevance": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Relevance",
            "description": "Graded relevance judgements; grade > 0 counts as relevant."
          },
          "ks": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "minItems": 1,
            "title": "Ks",
            "default": [
              1,
              5,
              10
            ]
          }
        },
        "type": "object",
        "required": [
          "ranked_ids",
          "relevance"
        ],
        "title": "RetrievalEvaluationRequest"
      },
      "RetrievalEvaluationResponse": {
        "properties": {
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/Metric"
            },
            "type": "array",
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "metrics"
        ],
        "title": "RetrievalEvaluationResponse"
      },
...
      "RetrievalMode": {
        "type": "string",
        "enum": [
          "manual",
          "adaptive"
        ],
        "title": "RetrievalMode"
      },
...
      "RetrievalRequest-Output": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query"
          },
          "top_k": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 10
          },
          "version": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Version",
            "description": "Corpus version; default current"
          },
          "strategy": {
            "$ref": "#/components/schemas/RetrievalStrategy",
            "default": "sparse"
          },
          "bm25": {
            "$ref": "#/components/schemas/Bm25Params-Output"
          },
          "hybrid": {
            "$ref": "#/components/schemas/HybridParams-Output"
          },
          "rerank": {
            "$ref": "#/components/schemas/RerankParams-Output"
          },
          "mode": {
            "$ref": "#/components/schemas/RetrievalMode",
            "description": "adaptive: the router replaces strategy, hybrid and rerank",
            "default": "manual"
          },
          "router": {
            "$ref": "#/components/schemas/RouterParams-Output"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query",
          "top_k",
          "version",
          "strategy",
          "bm25",
          "hybrid",
          "rerank",
          "mode",
          "router"
        ],
        "title": "RetrievalRequest"
      },
...
      "RetrievalStrategy": {
        "type": "string",
        "enum": [
          "sparse",
          "dense",
          "hybrid",
          "metadata",
          "graph",
          "multi_hop",
          "decomposition",
          "multimodal"
        ],
        "title": "RetrievalStrategy"
      },
...
      "RouteOption": {
        "type": "string",
        "enum": [
          "sparse",
          "dense",
          "hybrid_rrf",
          "hybrid_weighted"
        ],
        "title": "RouteOption"
      },
      "RouterDecision": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "query_id": {
            "type": "string",
            "title": "Query Id"
          },
          "policy": {
            "type": "string",
            "title": "Policy"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "policy_config_hash": {
            "type": "string",
            "title": "Policy Config Hash"
          },
          "analysis_hash": {
            "type": "string",
            "title": "Analysis Hash"
          },
          "option": {
            "$ref": "#/components/schemas/RouteOption",
            "description": "The selected retrieval option"
          },
          "preferred": {
            "$ref": "#/components/schemas/RouteOption",
            "description": "The policy's choice before availability"
          },
          "strategy": {
            "$ref": "#/components/schemas/RetrievalStrategy"
          },
          "hybrid": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HybridParams-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "rerank": {
            "$ref": "#/components/schemas/RerankParams-Output"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/RuleEvaluation"
            },
            "type": "array",
            "title": "Rules",
            "description": "Strategy rules in order, to the first match"
          },
          "rerank_rules": {
            "items": {
              "$ref": "#/components/schemas/RuleEvaluation"
            },
            "type": "array",
            "title": "Rerank Rules",
            "description": "Rerank rules, to the first match"
          },
          "alternatives": {
            "items": {
              "$ref": "#/components/schemas/RouteAlternative"
            },
            "type": "array",
            "title": "Alternatives"
          },
          "margin": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Margin",
            "description": "The deciding rule's margin; small = near a boundary"
          },
          "rationale": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rationale",
            "description": "The fired rules, filled in with measured values"
          },
          "configuration_hash": {
            "type": "string",
            "title": "Configuration Hash",
            "description": "Hash of the selected configuration, as the same manual request reports it"
          },
          "decision_hash": {
            "type": "string",
            "title": "Decision Hash",
            "description": "Hash of the decision, excluding ids and timestamps"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "query_id",
          "policy",
          "policy_version",
          "policy_config_hash",
          "analysis_hash",
          "option",
          "preferred",
          "strategy",
          "hybrid",
          "rerank",
          "rules",
          "rerank_rules",
          "alternatives",
          "margin",
          "rationale",
          "configuration_hash",
          "decision_hash",
          "created_at"
        ],
        "title": "RouterDecision"
      },
      "RouterParams-Input": {
        "properties": {
          "analyzer": {
            "type": "string",
            "title": "Analyzer",
            "description": "A registered query analyzer",
            "default": "heuristic"
          },
          "policy": {
            "type": "string",
            "title": "Policy",
            "description": "A registered router policy",
            "default": "rules-baseline"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RouterParams"
      },
      "RouterParams-Output": {
        "properties": {
          "analyzer": {
            "type": "string",
            "title": "Analyzer",
            "description": "A registered query analyzer",
            "default": "heuristic"
          },
          "policy": {
            "type": "string",
            "title": "Policy",
            "description": "A registered router policy",
            "default": "rules-baseline"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "analyzer",
          "policy"
        ],
        "title": "RouterParams"
      },
      "RouterRequest": {
        "properties": {
          "corpus_id": {
            "type": "string",
            "title": "Corpus Id"
          },
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Version",
            "description": "Corpus version; default current"
          },
          "top_k": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top K",
            "description": "The final top-k the route must serve",
            "default": 10
          },
          "router": {
            "$ref": "#/components/schemas/RouterParams-Input"
          }
        },
        "type": "object",
        "required": [
          "corpus_id",
          "query"
        ],
        "title": "RouterRequest"
      },
      "RouterResponse": {
        "properties": {
          "query": {
            "$ref": "#/components/schemas/Query"
          },
          "request": {
            "$ref": "#/components/schemas/RetrievalRequest-Output",
            "description": "The manual request the router selected: sending it reproduces the route"
          },
          "routing": {
            "$ref": "#/components/schemas/RoutingProvenance"
          }
        },
        "type": "object",
        "required": [
          "query",
          "request",
          "routing"
        ],
        "title": "RouterResponse"
      },
      "RoutingConfiguration": {
        "properties": {
          "analyzer": {
            "type": "string",
            "title": "Analyzer"
          },
          "analyzer_version": {
            "type": "string",
            "title": "Analyzer Version"
          },
          "analyzer_config_hash": {
            "type": "string",
            "title": "Analyzer Config Hash"
          },
          "policy": {
            "type": "string",
            "title": "Policy"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "policy_config_hash": {
            "type": "string",
            "title": "Policy Config Hash"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "analyzer",
          "analyzer_version",
          "analyzer_config_hash",
          "policy",
          "policy_version",
          "policy_config_hash"
        ],
        "title": "RoutingConfiguration",
        "description": "The identity of the routing step: which analyzer and policy, at which versions."
      },
      "RoutingProvenance": {
        "properties": {
          "routing": {
            "$ref": "#/components/schemas/RoutingConfiguration"
          },
          "routing_hash": {
            "type": "string",
            "title": "Routing Hash"
          },
          "analysis": {
            "$ref": "#/components/schemas/QueryAnalysis"
          },
          "decision": {
            "$ref": "#/components/schemas/RouterDecision"
          },
          "analysis_ms": {
            "type": "number",
            "title": "Analysis Ms",
            "description": "Query analysis, including corpus term statistics"
          },
          "decision_ms": {
            "type": "number",
            "title": "Decision Ms"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "routing",
          "routing_hash",
          "analysis",
          "decision",
          "analysis_ms",
          "decision_ms"
        ],
        "title": "RoutingProvenance"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}